+++ programs/master-token-program/src/instructions/investor/claim.rs
@@ -77,7 +77,7 @@ pub fn investor_claim_tokens<'info>(
             cpi_accounts,
             signer_seeds,
         );
-        token_2022::transfer_checked(cpi_ctx, total_tokens, SBT_DECIMALS as u8)?;
+        token_2022::transfer_checked(cpi_ctx, total_tokens, SBT_DECIMALS)?;
     }
     investor.months_claimed += total_months;
     ctx.accounts.category.allocated_unclaimed_tokens -= total_tokens;
